// Lec18/src/Types.tsx
export interface GitHubUser{
    login: string;
    avatar_url: string;
    name: string | null;
    company: string | null;
    blog: string | null;
    location: string | null;
    bio: string | null;
    public_repos: number;
    followers: number;
    following: number;
    created_at: string;
    twitter_username: string | null;
}